+++ zcash_client_backend/src/data_api/wallet.rs
@@ -1512,7 +1512,8 @@ where
             sapling_anchor,
             orchard_anchor,
             ironwood_anchor,
-            orchard_pool_bundle_type,
+            orchard_bundle_type: orchard_pool_bundle_type,
+            ironwood_bundle_type: orchard_pool_bundle_type,
         },
     );