+++ zcash_client_sqlite/src/wallet/orchard.rs
@@ -2096,7 +2096,8 @@ pub(crate) mod tests {
                     sapling_anchor: None,
                     orchard_anchor: None,
                     ironwood_anchor: Some(anchor),
-                    orchard_pool_bundle_type: orchard::builder::BundleType::DEFAULT,
+                    orchard_bundle_type: orchard::builder::BundleType::DEFAULT,
+                    ironwood_bundle_type: orchard::builder::BundleType::DEFAULT,
                 },
             );
             builder